assets/css/dc.css - adjust styles for small screens
Make banner and square-image feed captions smaller when screens get
below 575px

I've put a media query style in which shrinks square-image feed
captions to 1.3rem when the screen is < 574px (seems to be a standard
threshold in this theme)

It was clipping on the Land Explorer tool item because of the size of
the word "development" and possibly "LandExplorer".  1.3rem chosen
because I gather about 7% of screens can be 320px, and this just about
works (although longer words may tip it over the edge).